requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Chemistry
rdkit>=2022.9.1
deepchem>=2.7.0

# Equivariant networks
e3nn>=0.5.0
torch-cluster>=1.6.0

# Protein processing
fair-esm>=2.0.0
biopython>=1.81

# Normalizing flows (for density estimation)
nflows>=0.14

# Uncertainty quantification
mapie>=0.6.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Molecular fingerprints and descriptors
mordred>=1.2.0

# Optimization
optuna>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
py3Dmol>=2.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
hydra-core>=1.3.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Molecular generation baselines (for comparison)
guacamol>=0.5.4
